requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyMuPDF==1.23.8
openai==1.10.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0 